src/requirements.txt
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.2
PyPDF2==3.0.1         # pdfplumberの依存
pillow==10.2.0         # PDF内画像処理（pdfplumber依存）
